speech-to-task-agent/requirements.txt
faster-whisper
langchain
langchain-core
langchain-ollama
python-dotenv
